refactor(account-sites): remove onboarding site type shim
