--- src/services/accountSiteOnboarding/siteTypes.ts
@@ -1,12 +1,0 @@
-export {
-  ACCOUNT_SITE_TYPES,
-  ACCOUNT_SITE_TYPE_VALUES,
-  AIHUBMIX_API_ORIGIN,
-  AIHUBMIX_HOSTNAMES,
-  AIHUBMIX_LOGIN_PATH,
-  AIHUBMIX_WEB_ORIGIN,
-  MANAGED_SITE_TYPES,
-  SITE_TYPES,
-  type AccountSiteType,
-  type ManagedSiteType,
-} from "~/services/accountSiteDefinitions/siteTypes"
